requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.14
django-allauth==0.63.0
django-ckeditor==6.1.0
django-cors-headers==4.3.1
django-filter==24.2
django-guardian==3.0.0
django-js-asset==3.1.2
djangorestframework==3.15.2
drf-nested-routers==0.93.5
drf-yasg==1.21.8
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.3
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.11.0
